Refactor SubmitComment
